large changeset - isolate lr-finder
